% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{rmd_checkbox}
\alias{rmd_checkbox}
\title{Add checkbox}
\usage{
rmd_checkbox(choices, selected = NULL, label = NULL, inline = T,
  label_inline = T, format = "latex")
}
\arguments{
\item{choices}{A character vector for all the choices}

\item{selected}{A numeric vector for the index of all selected/checked items.}

\item{label}{Text string for label row. Default NULL.}

\item{inline}{T/F for whether choices should be aligned inline. Default T.}

\item{label_inline}{T/F for whether the label row should be placed inline
with the choices. Default T. If label_inline is set to be T, then inline
will be set to T as well.}

\item{format}{"latex" or "html". Controlled by global option
\code{rmdWidgets.format}}
}
\description{
Add checkbox
}
