% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textbox.R
\name{rmd_textbox}
\alias{rmd_textbox}
\title{Add textbox}
\usage{
rmd_textbox(text, textcolor = "black", title = "",
  titlecolor = "black", titleback = "yellow", backcolor = "white",
  rounded = "1", format = "latex")
}
\arguments{
\item{text}{Content of the text box.}

\item{textcolor}{Color of the text. Default black.}

\item{title}{Title of the text box. Default empty.}

\item{titlecolor}{Color of the title. Default white.}

\item{titleback}{Color of the background of the title. Default black.}

\item{backcolor}{Color of the background of the text. Default white.}

\item{rounded}{Describes how much the corner of the text box would be rounded.}

\item{format}{"latex" or "html". Controlled by global option
\code{rmdWidgets.format}}
}
\description{
Add textbox
}
